Added code to check object property value.
The Skills object is an extended Firebase
Object that syncs a collection of skills,
which are objects by themselves and have
a seniority property whose value is a
Firebase reference key from a seniority
reference record.

To be able to display the seniority name
next to each skill on any given place,
it's necessary to fetch each seniority
reference child by this value.

Sometimes it might happen that on the
client side there is code wich modifies
this model changing that value to an
object, further setting the old value
to an id property inside this newly
created object.

This changes helps the application to
check if the value associated with any
skill's seniority property is either a
UID string or an object.